feat: show and follow user location, reorganize component
